Refactor gbif functions, add admin action to update taxon
