+++ bims/models/taxonomy.py
@@ -389,3 +389,73 @@ class TaxonImage(models.Model):
     taxonomy = models.ForeignKey(
         Taxonomy, on_delete=models.CASCADE
     )
+
+
+def check_taxa_duplicates(taxon_name, taxon_rank):
+    """
+    Check for taxa duplicates, then merge if found
+    :param taxon_name: Name of the taxon to check
+    :param taxon_rank: Rank of the taxon to check
+    :return: Merged taxonomy
+    """
+    from django.db.models import Q, Min
+    from bims.utils.gbif import get_species
+    from bims.utils.fetch_gbif import merge_taxa_data
+
+    taxon_rank = taxon_rank.strip().upper()
+    taxon_name = taxon_name.strip()
+    taxa = Taxonomy.objects.filter(
+        Q(canonical_name__iexact=taxon_name) |
+        Q(legacy_canonical_name__icontains=taxon_name),
+        rank=taxon_rank
+    )
+    if not taxa.count() > 1:
+        return
+    preferred_taxa = taxa
+    accepted_taxa = taxa.filter(taxonomic_status='ACCEPTED')
+    if accepted_taxa.exists():
+        preferred_taxa = accepted_taxa
+    preferred_taxon = preferred_taxa.values('gbif_key', 'id').annotate(
+        Min('gbif_key')).order_by('gbif_key')[0]
+    preferred_taxon_gbif_data = get_species(preferred_taxon['gbif_key'])
+    preferred_taxon = Taxonomy.objects.get(
+        id=preferred_taxon['id']
+    )
+    for taxon in taxa[1:]:
+        gbif_data = get_species(taxon.gbif_key)
+        if not preferred_taxon_gbif_data:
+            preferred_taxon = taxon
+            preferred_taxon_gbif_data = gbif_data
+            continue
+        if not gbif_data:
+            continue
+        if gbif_data['taxonomicStatus'] == 'ACCEPTED':
+            preferred_taxon_gbif_data = gbif_data
+            preferred_taxon = taxon
+            continue
+        if 'issues' in gbif_data and len(gbif_data['issues']) > 0:
+            continue
+        if 'nubKey' not in gbif_data:
+            continue
+        if (
+            'taxonomicStatus' in gbif_data and
+            gbif_data['taxonomicStatus'] != 'ACCEPTED'
+        ):
+            continue
+        if 'key' not in preferred_taxon_gbif_data:
+            preferred_taxon = taxon
+            preferred_taxon_gbif_data = gbif_data
+            continue
+        if (
+            'key' in gbif_data and
+            gbif_data['key'] > preferred_taxon_gbif_data['key']
+        ):
+            continue
+        preferred_taxon = taxon
+        preferred_taxon_gbif_data = gbif_data
+
+    merge_taxa_data(
+        taxa_list=taxa.exclude(id=preferred_taxon.id),
+        excluded_taxon=preferred_taxon
+    )
+    return preferred_taxon
