+++ bims/utils/fetch_gbif.py
@@ -1,7 +1,5 @@
 import logging
-import json
 from django.db.models.fields.related import ForeignObjectRel
-from django.db.models import Q, Min
 from bims.utils.gbif import (
     get_children, find_species, get_species, get_vernacular_names,
     gbif_name_suggest
@@ -76,72 +74,6 @@ def merge_taxa_data(gbif_key='', excluded_taxon=None, taxa_list=None):
 
     if vernacular_names:
         excluded_taxon.vernacular_names.add(*vernacular_names)
-
-
-def check_taxa_duplicates(taxon_name, taxon_rank):
-    """
-    Check for taxa duplicates, then merge if found
-    :param taxon_name: Name of the taxon to check
-    :param taxon_rank: Rank of the taxon to check
-    :return: Merged taxonomy
-    """
-    taxon_rank = taxon_rank.strip().upper()
-    taxon_name = taxon_name.strip()
-    taxa = Taxonomy.objects.filter(
-        Q(canonical_name__iexact=taxon_name) |
-        Q(legacy_canonical_name__icontains=taxon_name),
-        rank=taxon_rank
-    )
-    if not taxa.count() > 1:
-        return
-    preferred_taxa = taxa
-    accepted_taxa = taxa.filter(taxonomic_status='ACCEPTED')
-    if accepted_taxa.exists():
-        preferred_taxa = accepted_taxa
-    preferred_taxon = preferred_taxa.values('gbif_key', 'id').annotate(
-        Min('gbif_key')).order_by('gbif_key')[0]
-    preferred_taxon_gbif_data = get_species(preferred_taxon['gbif_key'])
-    preferred_taxon = Taxonomy.objects.get(
-        id=preferred_taxon['id']
-    )
-    for taxon in taxa[1:]:
-        gbif_data = get_species(taxon.gbif_key)
-        if not preferred_taxon_gbif_data:
-            preferred_taxon = taxon
-            preferred_taxon_gbif_data = gbif_data
-            continue
-        if not gbif_data:
-            continue
-        if gbif_data['taxonomicStatus'] == 'ACCEPTED':
-            preferred_taxon_gbif_data = gbif_data
-            preferred_taxon = taxon
-            continue
-        if 'issues' in gbif_data and len(gbif_data['issues']) > 0:
-            continue
-        if 'nubKey' not in gbif_data:
-            continue
-        if (
-            'taxonomicStatus' in gbif_data and
-            gbif_data['taxonomicStatus'] != 'ACCEPTED'
-        ):
-            continue
-        if 'key' not in preferred_taxon_gbif_data:
-            preferred_taxon = taxon
-            preferred_taxon_gbif_data = gbif_data
-            continue
-        if (
-            'key' in gbif_data and
-            gbif_data['key'] > preferred_taxon_gbif_data['key']
-        ):
-            continue
-        preferred_taxon = taxon
-        preferred_taxon_gbif_data = gbif_data
-
-    merge_taxa_data(
-        taxa_list=taxa.exclude(id=preferred_taxon.id),
-        excluded_taxon=preferred_taxon
-    )
-    return preferred_taxon
 
 
 def create_or_update_taxonomy(
@@ -227,14 +159,17 @@ def create_or_update_taxonomy(
                 try:
                     vernacular_name, status = (
                         VernacularName.objects.get_or_create(
-                            name=result['vernacularName'],
-                            **fields
+                            name=result['vernacularName']
                         ))
                 except VernacularName.MultipleObjectsReturned:
                     vernacular_name = VernacularName.objects.filter(
-                        name=result['vernacularName'],
-                        **fields
+                        name=result['vernacularName']
                     ).first()
+
+                VernacularName.objects.filter(
+                    name__iexact=vernacular_name.name
+                ).update(**fields)
+
                 taxonomy.vernacular_names.add(vernacular_name)
     taxonomy.save()
     return taxonomy
@@ -245,17 +180,17 @@ def fetch_all_species_from_gbif(
     taxonomic_rank=None,
     gbif_key=None,
     parent=None,
-    should_get_children=False,
+    fetch_children=False,
     fetch_vernacular_names=False,
     use_name_lookup=True,
     **classifier):
     """
-    Get species detail and all species lower rank
+    Get species detail and all lower rank species
     :param species: species name
     :param taxonomic_rank: taxonomy rank e.g. class
     :param gbif_key: gbif key
     :param parent: taxonomy parent
-    :param should_get_children: fetch children or not
+    :param fetch_children: fetch children or not
     :param fetch_vernacular_names: fetch vernacular names or not
     :param use_name_lookup: use name_lookup to search species
     :return:
@@ -344,13 +279,13 @@ def fetch_all_species_from_gbif(
                 parent_taxonomy = fetch_all_species_from_gbif(
                     gbif_key=parent_key,
                     parent=None,
-                    should_get_children=False
+                    fetch_children=False
                 )
             if parent_taxonomy:
                 taxonomy.parent = parent_taxonomy
                 taxonomy.save()
 
-    # Check if there is accepted key
+    # Check if there is an accepted key
     if (
         'acceptedKey' in species_data and
         species_data['taxonomicStatus'] == 'SYNONYM'
@@ -358,13 +293,16 @@ def fetch_all_species_from_gbif(
         accepted_taxonomy = fetch_all_species_from_gbif(
             gbif_key=species_data['acceptedKey'],
             parent=taxonomy.parent,
-            should_get_children=False
+            fetch_children=False
         )
         if accepted_taxonomy:
+            if taxonomy.iucn_status:
+                accepted_taxonomy.iucn_status = taxonomy.iucn_status
+                accepted_taxonomy.save()
             taxonomy.accepted_taxonomy = accepted_taxonomy
             taxonomy.save()
 
-    if not should_get_children:
+    if not fetch_children:
         if taxonomy.legacy_canonical_name:
             legacy_canonical_name = taxonomy.legacy_canonical_name
             if legacy_name not in legacy_canonical_name:
@@ -391,46 +329,3 @@ def fetch_all_species_from_gbif(
                 parent=taxonomy
             )
         return taxonomy
-
-
-def check_taxon_parent(taxonomy):
-    """
-    Check if parent of the taxonomy is identical with one from gbif
-    :param taxonomy: Taxonomy object
-    :return: taxonomy
-    """
-    logger.debug('Check taxon parent')
-    if not taxonomy.gbif_key:
-        logger.debug('Gbif key not found')
-        return
-    gbif_data = get_species(taxonomy.gbif_key)
-
-    parent_key_found = True
-    fetch_parent = False
-
-    if 'parentKey' not in gbif_data:
-        logger.debug('Missing parent key')
-        parent_key_found = False
-        fetch_parent = True
-
-    current_parent = taxonomy.parent
-    if current_parent and parent_key_found:
-        current_parent_gbif_key = current_parent.gbif_key
-        if current_parent_gbif_key != int(gbif_data['parentKey']):
-            fetch_parent = True
-        else:
-            logger.debug('Taxon parent is correct')
-
-    if fetch_parent:
-        if parent_key_found:
-            logger.debug('Updating parent')
-            parents = Taxonomy.objects.filter(gbif_key=gbif_data['parentKey'])
-            if parents.exists():
-                taxonomy.parent = parents[0]
-            else:
-                parent = fetch_all_species_from_gbif(
-                    gbif_key=gbif_data['parentKey']
-                )
-                taxonomy.parent = parent
-            taxonomy.save()
-    return taxonomy
